+++ src/components/alldecks/AllDecksList.js
@@ -4,7 +4,7 @@ import { PlayerClassContext } from "../deckbuilder/playerclass/PlayerClassProvid
 import "./AllDeckList.css"
 import { RatingContext } from "../rating/RatingProvider"
 import { MDBDataTable} from 'mdbreact';
-import { useHistory } from "react-router-dom"
+import { useHistory, Link } from "react-router-dom"
 import { UserContext } from '../user/UserProvider'
 
 export const AllDeckList = () => {
@@ -82,7 +82,7 @@ export const AllDeckList = () => {
 
                 let obj = {
                     rating: theRating,
-                    deckname: <div onClick={goToDeck}>{deck.deck_name}</div>,
+                    deckname: <Link to={`/decks/${deck.id}`}>{deck.deck_name}</Link>,
                     class: theClass?.name,
                     published: new Date(deck.published).toLocaleDateString('en-US'),
                     author: deckAuthor?.username
