+++ src/components/deckbuilder/cardoption/CardByNeutralClassCard.js
@@ -41,8 +41,6 @@ export const CardByNeutralClassCard = ({card}) => {
 
         if (card.rarity !== "LEGENDARY" && countCards < 2){
             if (cardCountForDecks < 30) {
-                // const perCardCount = ++countCards
-                // setCountCards(perCardCount)
                 
                 const cardFinder = localCards.find(c => c.dbfId === card.dbfId)
                 const carddbfId = cardFinder.dbfId
@@ -55,18 +53,10 @@ export const CardByNeutralClassCard = ({card}) => {
                 }
 
                 updateDeckCart(deckCartObj)
-                // if(countCards === 2){
-                //     let theCount = document.getElementById(`${card.dbfId}`)
-                //     let theX = document.getElementById(`x--${card.dbfId}`)
-                //     theCount.classList.add("greyscale")
-                //     theX.classList.remove('isVisible')
 
-                // }
             }
         } else if (card.rarity === "LEGENDARY" && countCards < 1){
             if (cardCountForDecks < 30) {
-                // const perCardCount = ++countCards
-                // setCountCards(perCardCount)
 
                 const cardFinder = localCards.find(c => c.dbfId === card.dbfId)
                 const carddbfId = cardFinder.dbfId
@@ -79,13 +69,6 @@ export const CardByNeutralClassCard = ({card}) => {
                 }
 
                 updateDeckCart(deckCartObj)
-                // if(countCards === 1){
-                //     let theCount = document.getElementById(`${card.dbfId}`)
-                //     let theX = document.getElementById(`x--${card.dbfId}`)
-                //     theCount.classList.add("greyscale")
-                //     theX.classList.remove('isVisible')
-
-                // }
             }
         }   
     }
@@ -95,7 +78,7 @@ export const CardByNeutralClassCard = ({card}) => {
         <section className="neutralCardViewerOptions">
               <div className="cardImage">
                 <img src={`https://art.hearthstonejson.com/v1/render/latest/enUS/256x/${card.id}.png`} className="card_image" id={`${card?.dbfId}`} onClick={event => cardWasClicked(card)} alt={`${card?.name}`}/>
-                <img src={'/images/redx.png'} className="redx isVisible" id={`x--${card.dbfId}`}/>
+                <img src={'/images/redx.png'} className="redx isVisible" id={`x--${card.dbfId}`} alt="Red X"/>
               </div>
               <div className="cardCount" id={`${card.name}`}>Added: {countCards}</div>
         </section>
