+++ src/components/deckview/DeckViewList.css
@@ -100,6 +100,6 @@
   display: none;
 }
 
-body {
-  /* color: white; */
+.white {
+  color: whitesmoke;
 }
